Refactored LocalApigwService by moving some event generation logic out
