+++ samcli/local/apigw/local_apigw_service.py
@@ -1,17 +1,13 @@
 """API Gateway Local Service"""
-# pylint: disable=too-many-lines
-# TODO (lucashuy): this module needs to be refactored
 
 import base64
 import io
 import json
 import logging
 from dataclasses import dataclass
-from datetime import datetime
-from time import time
 from typing import List, Optional
 
-from flask import Flask, request
+from flask import Flask, Request, request
 from werkzeug.datastructures import Headers
 from werkzeug.routing import BaseConverter
 from werkzeug.serving import WSGIRequestHandler
@@ -19,20 +15,12 @@ from werkzeug.serving import WSGIRequestHandler
 from samcli.commands.local.lib.exceptions import UnsupportedInlineCodeError
 from samcli.lib.providers.provider import Cors
 from samcli.lib.utils.stream_writer import StreamWriter
+from samcli.local.apigw.event_constructor import EventConstructor
 from samcli.local.apigw.exceptions import LambdaResponseParseException, PayloadFormatVersionValidateException
-from samcli.local.events.api_event import (
-    ApiGatewayLambdaEvent,
-    ApiGatewayV2LambdaEvent,
-    ContextHTTP,
-    ContextIdentity,
-    RequestContext,
-    RequestContextV2,
-)
+from samcli.local.apigw.path_converter import PathConverter
+from samcli.local.apigw.service_error_responses import ServiceErrorResponses
 from samcli.local.lambdafn.exceptions import FunctionNotFound
 from samcli.local.services.base_local_service import BaseLocalService, LambdaOutputParser
-
-from .path_converter import PathConverter
-from .service_error_responses import ServiceErrorResponses
 
 LOG = logging.getLogger(__name__)
 
@@ -315,6 +303,58 @@ class LocalApigwService(BaseLocalService):
         # Something went wrong
         self._app.register_error_handler(500, ServiceErrorResponses.lambda_failure_response)
 
+    def _generate_lambda_event(self, flask_request: Request, route: Route, method: str, endpoint: str) -> str:
+        """
+        Helper function to generate the correct Lambda event
+
+        Parameters
+        ----------
+        flask_request: Request
+            The global Flask Request object
+        route: Route
+            The Route that was called
+        method: str
+            The method of the request (eg. GET, POST) from the Flask request
+        endpoint: str
+            The endpoint of the request from the Flask request
+
+        Returns
+        -------
+        str
+            JSON string of event properties
+        """
+        # Default to None
+        # For Http Apis with payload version 1.0, API Gateway never sends the OperationName.
+        route_key = None
+
+        # TODO: Rewrite the logic below to use version 2.0 when an invalid value is provided
+        # the Lambda Event 2.0 is only used for the HTTP API gateway with defined payload format version equal 2.0
+        # or none, as the default value to be used is 2.0
+        # https://docs.aws.amazon.com/apigatewayv2/latest/api-reference/apis-apiid-integrations.html#apis-apiid-integrations-prop-createintegrationinput-payloadformatversion
+        if route.event_type == Route.HTTP and route.payload_format_version in [None, "2.0"]:
+            apigw_endpoint = PathConverter.convert_path_to_api_gateway(endpoint)
+            route_key = self._v2_route_key(method, apigw_endpoint, route.is_default_route)
+
+            return EventConstructor.v2_event_http(
+                flask_request=flask_request,
+                port=self.port,
+                binary_types=self.api.binary_media_types,
+                stage_name=self.api.stage_name,
+                stage_variables=self.api.stage_variables,
+                route_key=route_key,
+            )
+        elif route.event_type == Route.API:
+            route_key = route.operation_name
+
+        return EventConstructor.v1_event(
+            flask_request=flask_request,
+            port=self.port,
+            binary_types=self.api.binary_media_types,
+            stage_name=self.api.stage_name,
+            stage_variables=self.api.stage_variables,
+            operation_name=route_key,
+        )
+
     def _request_handler(self, **kwargs):
         """
         We handle all requests to the host:port. The general flow of handling a request is as follows
@@ -356,41 +396,7 @@ class LocalApigwService(BaseLocalService):
             return self.service_response("", headers, 200)
 
         try:
-            # TODO: Rewrite the logic below to use version 2.0 when an invalid value is provided
-            # the Lambda Event 2.0 is only used for the HTTP API gateway with defined payload format version equal 2.0
-            # or none, as the default value to be used is 2.0
-            # https://docs.aws.amazon.com/apigatewayv2/latest/api-reference/apis-apiid-integrations.html#apis-apiid-integrations-prop-createintegrationinput-payloadformatversion
-            if route.event_type == Route.HTTP and route.payload_format_version in [None, "2.0"]:
-                apigw_endpoint = PathConverter.convert_path_to_api_gateway(endpoint)
-                route_key = self._v2_route_key(method, apigw_endpoint, route.is_default_route)
-                event = self._construct_v_2_0_event_http(
-                    request,
-                    self.port,
-                    self.api.binary_media_types,
-                    self.api.stage_name,
-                    self.api.stage_variables,
-                    route_key,
-                )
-            elif route.event_type == Route.API:
-                # The OperationName is only sent to the Lambda Function from API Gateway V1(Rest API).
-                event = self._construct_v_1_0_event(
-                    request,
-                    self.port,
-                    self.api.binary_media_types,
-                    self.api.stage_name,
-                    self.api.stage_variables,
-                    route.operation_name,
-                )
-            else:
-                # For Http Apis with payload version 1.0, API Gateway never sends the OperationName.
-                event = self._construct_v_1_0_event(
-                    request,
-                    self.port,
-                    self.api.binary_media_types,
-                    self.api.stage_name,
-                    self.api.stage_variables,
-                    None,
-                )
+            event = self._generate_lambda_event(request, route, method, endpoint)
         except UnicodeDecodeError as error:
             LOG.error("UnicodeDecodeError while processing HTTP request: %s", error)
             return ServiceErrorResponses.lambda_failure_response()
@@ -709,311 +715,3 @@ class LocalApigwService(BaseLocalService):
             processed_headers.add(header, headers[header])
 
         return processed_headers
-
-    @staticmethod
-    def _construct_v_1_0_event(
-        flask_request, port, binary_types, stage_name=None, stage_variables=None, operation_name=None
-    ):
-        """
-        Helper method that constructs the Event to be passed to Lambda
-
-        :param request flask_request: Flask Request
-        :param port: the port number
-        :param binary_types: list of binary types
-        :param stage_name: Optional, the stage name string
-        :param stage_variables: Optional, API Gateway Stage Variables
-        :return: String representing the event
-        """
-        # pylint: disable-msg=too-many-locals
-
-        identity = ContextIdentity(source_ip=flask_request.remote_addr)
-
-        endpoint = PathConverter.convert_path_to_api_gateway(flask_request.endpoint)
-        method = flask_request.method
-        protocol = flask_request.environ.get("SERVER_PROTOCOL", "HTTP/1.1")
-        host = flask_request.host
-
-        request_data = flask_request.get_data()
-
-        request_mimetype = flask_request.mimetype
-
-        is_base_64 = LocalApigwService._should_base64_encode(binary_types, request_mimetype)
-
-        if is_base_64:
-            LOG.debug("Incoming Request seems to be binary. Base64 encoding the request data before sending to Lambda.")
-            request_data = base64.b64encode(request_data)
-
-        if request_data:
-            # Flask does not parse/decode the request data. We should do it ourselves
-            # Note(xinhol): here we change request_data's type from bytes to str and confused mypy
-            # We might want to consider to use a new variable here.
-            request_data = request_data.decode("utf-8")
-
-        query_string_dict, multi_value_query_string_dict = LocalApigwService._query_string_params(flask_request)
-
-        context = RequestContext(
-            resource_path=endpoint,
-            http_method=method,
-            stage=stage_name,
-            identity=identity,
-            path=endpoint,
-            protocol=protocol,
-            domain_name=host,
-            operation_name=operation_name,
-        )
-
-        headers_dict, multi_value_headers_dict = LocalApigwService._event_headers(flask_request, port)
-
-        event = ApiGatewayLambdaEvent(
-            http_method=method,
-            body=request_data,
-            resource=endpoint,
-            request_context=context,
-            query_string_params=query_string_dict,
-            multi_value_query_string_params=multi_value_query_string_dict,
-            headers=headers_dict,
-            multi_value_headers=multi_value_headers_dict,
-            path_parameters=flask_request.view_args,
-            path=flask_request.path,
-            is_base_64_encoded=is_base_64,
-            stage_variables=stage_variables,
-        )
-
-        event_str = json.dumps(event.to_dict(), sort_keys=True)
-        LOG.debug("Constructed String representation of Event to invoke Lambda. Event: %s", event_str)
-        return event_str
-
-    @staticmethod
-    def _construct_v_2_0_event_http(
-        flask_request,
-        port,
-        binary_types,
-        stage_name=None,
-        stage_variables=None,
-        route_key=None,
-        request_time_epoch=int(time()),
-        request_time=datetime.utcnow().strftime("%d/%b/%Y:%H:%M:%S +0000"),
-    ):
-        """
-        Helper method that constructs the Event 2.0 to be passed to Lambda
-
-        https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-develop-integrations-lambda.html
-
-        :param request flask_request: Flask Request
-        :param port: the port number
-        :param binary_types: list of binary types
-        :param stage_name: Optional, the stage name string
-        :param stage_variables: Optional, API Gateway Stage Variables
-        :param route_key: Optional, the route key for the route
-        :return: String representing the event
-        """
-        # pylint: disable-msg=too-many-locals
-        method = flask_request.method
-
-        request_data = flask_request.get_data()
-
-        request_mimetype = flask_request.mimetype
-
-        is_base_64 = LocalApigwService._should_base64_encode(binary_types, request_mimetype)
-
-        if is_base_64:
-            LOG.debug("Incoming Request seems to be binary. Base64 encoding the request data before sending to Lambda.")
-            request_data = base64.b64encode(request_data)
-
-        if request_data is not None:
-            # Flask does not parse/decode the request data. We should do it ourselves
-            request_data = request_data.decode("utf-8")
-
-        query_string_dict = LocalApigwService._query_string_params_v_2_0(flask_request)
-
-        cookies = LocalApigwService._event_http_cookies(flask_request)
-        headers = LocalApigwService._event_http_headers(flask_request, port)
-        context_http = ContextHTTP(method=method, path=flask_request.path, source_ip=flask_request.remote_addr)
-        context = RequestContextV2(
-            http=context_http,
-            route_key=route_key,
-            stage=stage_name,
-            request_time_epoch=request_time_epoch,
-            request_time=request_time,
-        )
-
-        event = ApiGatewayV2LambdaEvent(
-            route_key=route_key,
-            raw_path=flask_request.path,
-            raw_query_string=flask_request.query_string.decode("utf-8"),
-            cookies=cookies,
-            headers=headers,
-            query_string_params=query_string_dict,
-            request_context=context,
-            body=request_data,
-            path_parameters=flask_request.view_args,
-            is_base_64_encoded=is_base_64,
-            stage_variables=stage_variables,
-        )
-
-        event_str = json.dumps(event.to_dict())
-        LOG.debug("Constructed String representation of Event Version 2.0 to invoke Lambda. Event: %s", event_str)
-        return event_str
-
-    @staticmethod
-    def _query_string_params(flask_request):
-        """
-        Constructs an APIGW equivalent query string dictionary
-
-        Parameters
-        ----------
-        flask_request request
-            Request from Flask
-
-        Returns dict (str: str), dict (str: list of str)
-        -------
-            Empty dict if no query params where in the request otherwise returns a dictionary of key to value
-
-        """
-        query_string_dict = {}
-        multi_value_query_string_dict = {}
-
-        # Flask returns an ImmutableMultiDict so convert to a dictionary that becomes
-        # a dict(str: list) then iterate over
-        for query_string_key, query_string_list in flask_request.args.lists():
-            query_string_value_length = len(query_string_list)
-
-            # if the list is empty, default to empty string
-            if not query_string_value_length:
-                query_string_dict[query_string_key] = ""
-                multi_value_query_string_dict[query_string_key] = [""]
-            else:
-                query_string_dict[query_string_key] = query_string_list[-1]
-                multi_value_query_string_dict[query_string_key] = query_string_list
-
-        return query_string_dict, multi_value_query_string_dict
-
-    @staticmethod
-    def _query_string_params_v_2_0(flask_request):
-        """
-        Constructs an APIGW equivalent query string dictionary using the 2.0 format
-        https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-develop-integrations-lambda.html#2.0
-
-        Parameters
-        ----------
-        flask_request request
-            Request from Flask
-
-        Returns dict (str: str)
-        -------
-            Empty dict if no query params where in the request otherwise returns a dictionary of key to value
-
-        """
-        query_string_dict = {}
-
-        # Flask returns an ImmutableMultiDict so convert to a dictionary that becomes
-        # a dict(str: list) then iterate over
-        query_string_dict = {
-            query_string_key: ",".join(query_string_list)
-            for query_string_key, query_string_list in flask_request.args.lists()
-        }
-
-        return query_string_dict
-
-    @staticmethod
-    def _event_headers(flask_request, port):
-        """
-        Constructs an APIGW equivalent headers dictionary
-
-        Parameters
-        ----------
-        flask_request request
-            Request from Flask
-        int port
-            Forwarded Port
-        cors_headers dict
-            Dict of the Cors properties
-
-        Returns dict (str: str), dict (str: list of str)
-        -------
-            Returns a dictionary of key to list of strings
-
-        """
-        headers_dict = {}
-        multi_value_headers_dict = {}
-
-        # Multi-value request headers is not really supported by Flask.
-        # See https://github.com/pallets/flask/issues/850
-        for header_key in flask_request.headers.keys():
-            headers_dict[header_key] = flask_request.headers.get(header_key)
-            multi_value_headers_dict[header_key] = flask_request.headers.getlist(header_key)
-
-        headers_dict["X-Forwarded-Proto"] = flask_request.scheme
-        multi_value_headers_dict["X-Forwarded-Proto"] = [flask_request.scheme]
-
-        headers_dict["X-Forwarded-Port"] = str(port)
-        multi_value_headers_dict["X-Forwarded-Port"] = [str(port)]
-        return headers_dict, multi_value_headers_dict
-
-    @staticmethod
-    def _event_http_cookies(flask_request):
-        """
-        All cookie headers in the request are combined with commas.
-
-        https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-develop-integrations-lambda.html
-
-        Parameters
-        ----------
-        flask_request request
-            Request from Flask
-
-        Returns list
-        -------
-            Returns a list of cookies
-
-        """
-        cookies = []
-        for cookie_key in flask_request.cookies.keys():
-            cookies.append("{}={}".format(cookie_key, flask_request.cookies.get(cookie_key)))
-        return cookies
-
-    @staticmethod
-    def _event_http_headers(flask_request, port):
-        """
-        Duplicate headers are combined with commas.
-
-        https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-develop-integrations-lambda.html
-
-        Parameters
-        ----------
-        flask_request request
-            Request from Flask
-
-        Returns list
-        -------
-            Returns a list of cookies
-
-        """
-        headers = {}
-        # Multi-value request headers is not really supported by Flask.
-        # See https://github.com/pallets/flask/issues/850
-        for header_key in flask_request.headers.keys():
-            headers[header_key] = flask_request.headers.get(header_key)
-
-        headers["X-Forwarded-Proto"] = flask_request.scheme
-        headers["X-Forwarded-Port"] = str(port)
-        return headers
-
-    @staticmethod
-    def _should_base64_encode(binary_types, request_mimetype):
-        """
-        Whether or not to encode the data from the request to Base64
-
-        Parameters
-        ----------
-        binary_types list(basestring)
-            Corresponds to self.binary_types (aka. what is parsed from SAM Template
-        request_mimetype str
-            Mimetype for the request
-
-        Returns
-        -------
-            True if the data should be encoded to Base64 otherwise False
-
-        """
-        return request_mimetype in binary_types or "*/*" in binary_types
